fix: ensure singleton container
