+++ src/dbt_core_interface/dbt_templater/__init__.py
@@ -1,6 +1,8 @@
 # pyright: reportAny=false
 """Defines the hook endpoints for the dbt templater plugin."""
 
+from __future__ import annotations
+
 import typing as t
 
 try:
